requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
redis==5.0.1
qdrant-client>=1.12.0
httpx==0.26.0
pydantic-settings==2.1.0
websockets==12.0
python-dotenv==1.0.0
